Move standard entity failures under `entity.failure` package
