+++ server/src/main/proto/spine/server/entity/failures.proto
@@ -23,8 +23,7 @@ package spine.server.entity;
 
 option (type_url_prefix) = "type.spine3.org";
 option (SPI_all) = true;
-//TODO:2017-02-21:alexander.yevsyukov: Have another package.
-option java_package = "org.spine3.server.entity.status";
+option java_package = "org.spine3.server.entity.failure";
 // Keep all the failures under the same outer class called `Failures`.
 option java_generate_equals_and_hash = true;
