+++ server/src/test/java/org/spine3/server/entity/VisibilityTests.java
@@ -23,8 +23,8 @@ package org.spine3.server.entity;
 import com.google.protobuf.StringValue;
 import org.junit.Before;
 import org.junit.Test;
-import org.spine3.server.entity.status.CannotModifyArchivedEntity;
-import org.spine3.server.entity.status.CannotModifyDeletedEntity;
+import org.spine3.server.entity.failure.CannotModifyArchivedEntity;
+import org.spine3.server.entity.failure.CannotModifyDeletedEntity;
 
 import java.util.concurrent.ThreadLocalRandom;
